Break mds-metrics-sheet out into its own package, and convert to TS.
