+++ package.json
@@ -22,6 +22,7 @@
     "@types/pg": "7.4.14",
     "@types/point-in-polygon": "1.0.0",
     "@types/redis": "2.8.13",
+    "@types/request-promise": "4.1.44",
     "@types/sinon": "7.0.13",
     "@types/supertest": "2.0.7",
     "@types/uuid": "3.4.5",
