--- packages/mds-daily/tools/metrics-sheet/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "mds-metrics-sheet",
-  "version": "0.0.1",
-  "description": "Automate reporting api insert into LADOT compliance spreadsheet",
-  "main": "index.js",
-  "private": true,
-  "repository": {
-    "type": "git",
-    "url": ""
-  },
-  "engines": {
-    "node": ">=8"
-  },
-  "scripts": {
-    "start": "node ./index.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "City of Los Angeles",
-  "license": "Apache-2.0",
-  "dependencies": {
-    "dotenv": "^7.0.0",
-    "google-spreadsheet": "^2.0.7",
-    "request": "^2.88.0",
-    "request-promise": "^4.2.4"
-  }
-}
