+++ packages/mds-metrics-sheet/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "mds-metrics-sheet",
+  "version": "0.0.1",
+  "description": "Automate reporting api insert into LADOT compliance spreadsheet",
+  "private": true,
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "scripts": {
+    "start": "node ./dist/geo-metrics.js",
+    "bundle": "webpack --mode=production --env.npm_package_name=${npm_package_name} --env.npm_package_version=${npm_package_version}",
+    "build": "tsc --build .",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "City of Los Angeles",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "dotenv": "8.0.0",
+    "google-spreadsheet": "2.0.7",
+    "mds-logger": "0.1.0",
+    "mds-providers": "0.1.0",
+    "request": "2.88.0",
+    "request-promise": "4.2.4"
+  }
+}
