+++ packages/mds-metrics-sheet/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": "../../tsconfig.settings.json",
+  "compilerOptions": {
+    "outDir": "dist"
+  },
+  "strict": false,
+  "references": [{ "path": "../../packages/mds-logger" }, { "path": "../../packages/mds-providers" }]
+}
